+++ RELEASE-NOTES.md
@@ -1,13 +1,19 @@
 ## RELEASE NOTES
+### Version 6.16.0-follow-up-query-check-your-answers-v4
+**EUI-8607** Query management follow up query check your answers page changes
+
+### Version 6.16.0-query-details-ui-changes-v3
+**EUI-8604** Query details UI changes on the Query tab
+
+## Version 6.16.0-query-management-raise-a-query-page-changes
+**EUI-8602** Query management raise a query page changes
+
 ### Version 6.16-hotfix-EUI-8515-case-flags-submission
 **EUI-8515** Fix Case Flags and Linked Cases submissions not to depend on presence of "Check your answers" page
 **EUI-8550** Fix incorrect behaviour of showEventNotes() function on Case Event submission, introduced in error by EUI-6693
 
 ### Version 6.16.0-raise-a-query-check-your-answers-v2
 **EUI-8603** Query management raise a query check your answers page changes
-
-### Version 6.16.0-follow-up-query-check-your-answers
-**EUI-8607** Query management follow up query check your answers page changes
 
 ### Version 6.16.0-create-qualifying-questions-component-step-2
 **EUI-8524** Create qualifying questions component step 2